Adjust v4l2 import in prep for its separation
v4l2 is a separate python package.  We should not be replicating its
source code here.  Just as with other python libraries, we should
require it to be installed on the system we're working on.

When v4l2 is installed on a system, it is imported with:

	import v4l2

However, we were importing it with:

	from v4l2 import v4l2

Which means that when we remove v4l2 from our source, the import will
no longer work.  This patch adjusts the __init__.py in our copy of
v4l2 so that the import is the same as if it were installed on the
system.