Allow find_markers to take opencv's python images
opencv provides a python API.  Its API is provided by a python module
written in C.  That python library passes IplImages around wrapped in
Python's object system, so we have to extract them from this system in
order to use them.  This patch augments find_markers and
find_markers_fp so that they can either take one of libkoki's
IplImages or an image from opencv's python library.

This patch covers the opencv to libkoki direction for IplImages.
Going the other way (libkoki to opencv) is going to be more tricky, as
it will require us to inject it into a suitable PyObject and get its
data managed (or potentially not managed!) by the python object
system.